refactor: mobile menu
