--- app/components/(functional)/MobileAccordion/MobileAccordion.module.css
@@ -1,51 +1,0 @@
-.accordionTitle {
-    width: auto;
-    height: 55px;
-    font-size: 18px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 4px;
-    background-color: #00150f;
-    border: 0;
-    color: white;
-    cursor: pointer;
-    border-top: 1px solid rgba(255, 215, 0, 0.5);
-    border-bottom: 1px solid rgba(255, 215, 0, 0.5);
-  }
-  
-  .accordionContent {
-    padding: 0 24px;
-    font-size: 18px;
-  }
-  
-  .accordionCaret {
-    width: 24px;
-    height: 24px;
-  }
-  
-  .accordionDropdownButton {
-    padding: 1rem;
-  }
-  
-  @media screen and (max-width: 1025px) {
-    .accordionTitle {
-      display: flex;
-      justify-content: space-between;
-      width: 100%;
-      position: relative;
-      padding: 0 24px;
-      color: white;
-    }
-  }
-  
-  @media screen and (max-width: 768px) {
-    .accordionTitle {
-      display: flex;
-      justify-content: space-between;
-      width: 100%;
-      position: relative;
-      padding: 0 24px;
-      color: white;
-    }
-  }
