+++ app/components/index.ts
@@ -24,8 +24,8 @@ import { notify } from "./(display)/Notification/Notification";
 import Filters from "./(functional)/Filters/Filters";
 import LatestJobCard from "./(functional)/LatestJobCard/LatestJobCard";
 import JobCard from "./(functional)/JobCard/JobCard";
-import MobileMenu from "./(functional)/MobileMenu/MobileMenu";
-import MobileAccordion from "./(functional)/MobileAccordion/MobileAccordion";
+import MobileMenu from "./(functional)/MobileMenu";
+import MobileAccordion from "./(functional)/MobileAccordion";
 import MenuItem from "./(functional)/MenuItem/MenuItem";
 import Search from "./(functional)/Search/Search";
 import Login from "./(functional)/Login/Login";
